+++ bottom-bar/src/main/res/layout/bb_bottom_bar_item_container.xml
@@ -5,7 +5,6 @@
         android:id="@+id/bb_bottom_bar_shadow"
         android:layout_width="match_parent"
         android:layout_height="4dp"
-        android:layout_gravity="bottom"
         android:src="@drawable/bb_bottom_bar_top_shadow"/>
 
     <FrameLayout
